requirements.txt
Flask>=3.0.0
stripe>=7.0.0
python-dotenv>=1.0.0

